backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]>=2.0.40
aiosqlite==0.20.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
# passlib removed - using hashlib for Python 3.14 compatibility
python-multipart==0.0.18
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiofiles==24.1.0
